+++ code/Meerkat.Security.Hmac.Test/Net/Http/Md5Fixture.cs
@@ -1,7 +1,9 @@
 ﻿using System;
 using System.Net.Http;
 using System.Threading.Tasks;
+
 using Meerkat.Net.Http;
+
 using NUnit.Framework;
 
 namespace Meerkat.Test.Net.Http