+++ code/Meerkat.Security.Hmac.WebApi/Owin/Security/Infrastructure/HmacAuthenticationOptions.cs
@@ -1,0 +1,13 @@
+﻿using Meerkat.Security;
+using Microsoft.Owin.Security;
+
+namespace Meerkat.Owin.Security.Infrastructure
+{
+    public class HmacAuthenticationOptions : AuthenticationOptions
+    {
+        public HmacAuthenticationOptions() : base(HmacAuthentication.AuthenticationScheme)
+        {
+            AuthenticationMode = AuthenticationMode.Active;
+        }
+    }
+}