+++ code/Sample.Web/App_Start/WebApiConfig.cs
@@ -1,7 +1,10 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
+﻿using System.Linq;
 using System.Web.Http;
+using System.Web.Http.Filters;
+
+using Microsoft.Practices.Unity.WebApi;
+
+using Sample.Web.Http;
 
 namespace Sample.Web
 {
@@ -9,10 +12,26 @@ namespace Sample.Web
     {
         public static void Register(HttpConfiguration config)
         {
+            // Wire up dependency resolution for WebAPI
+            var container = UnityConfig.GetConfiguredContainer();
+            config.DependencyResolver = new UnityHierarchicalDependencyResolver(container);
+
             // Web API configuration and services
+
+            // Use Unity filter provider to get attribute injection
+            var providers = config.Services.GetFilterProviders().ToList();
+            var defaultProvider = providers.Single(x => x is ActionDescriptorFilterProvider);
+            config.Services.Remove(typeof(IFilterProvider), defaultProvider);
+            config.Services.Add(typeof(IFilterProvider), new UnityFilterProvider(container));
 
             // Web API routes
             config.MapHttpAttributeRoutes();
+
+            config.Routes.MapHttpRoute(
+                     name: "WithActionApi",
+                     routeTemplate: "api/{controller}/{action}/{id}",
+                     defaults: new { id = System.Web.Http.RouteParameter.Optional }
+                     );
 
             config.Routes.MapHttpRoute(
                 name: "DefaultApi",